Introduce NewClient constructor
This change makes GetProfile side-effect-free, and gives callers a
constructor (NewClient) for initializing a DPE client. Because the
client needs to know what profile the implementation thinks it
implemented, the client always needs to call this function. Exposing a
constructor allows users to avoid accidentally using invalid state
(e.g., by creating a Client struct and not calling GetProfile).

This change renames DPEClient to just 'Client' since DPE is implied in
the package.