Extend Platform trait and remove locality array from DPE instance
This commit extends the platform trait so that the implementation
can define the vendor_id, vendor_sku, and auto_init_locality. We
also remove localities from DpeInstance. This means that DPE
no longer knows about all the available localities at startup, and
it now just trusts that each command is called with a valid locality.