Make DPE cmd execution return the response struct with a response hdr
Instead of serializing the DPE response, this change makes DPE
command execution return the response struct itself with a
response header, so that the caller can decide what to do with it.
We need to add the header to each DPE response struct so that
if and when we serialize it using as_bytes, the header exists in
the serialized buffer.