Refactor Hasher
The ownership of the Hasher interface was very difficult to implement in
practice. In particular, if the hasher needs to hold references to
shared data, the lifetime of the returned hasher if very complicated.

Update this interface to return a HashCtx. The idea is that this can
hold data, but does not need to own the actual hasher interface (which
may be a hardware interface).