Make alias cert match crypto algorithm
This change adds a test cert created with a p384 key and makes
platform/crypto run based on the algorithm being used. Also,
we make TEST_ISSUER equal to the subject name of TEST_CERT_CHAIN
so cert validation can work.

Additionally, we refactor new_for_test to only take in command ids
instead of the whole command. Earlier, this function was confusing
since we take in the entire command but don't do anything with it
other than get the command id, which gives the impression that we
are serializing the command too, which is incorrect.