Remove usage of unsafe Rust in CryptoBuf
Previously, we were unsafely using ArrayVec::set_len in the
CryptoBuf implementation. We were also not checking that the
size of the input slice bytes == algs.size() which resulted
in partially initialized CryptoBufs. This change fixes both
of these issues.

Now, there are no more usages of unsafe Rust in DPE.