Set internal input flags on child contexts instead of parent
Previously we were setting the internal_dice and internal_info
flags on the parent context in derive child. This is a bug since
attestation could fail under the following scenario:

1. Assume context[0] is auto-initialized to the default context
2. Call DeriveChild with retain parent
3. Call Sign on the default context (context[0])
4. Call DeriveChild with retain parent and internal_input_info.
5. Call CertifyKey on the default context (context[0])

Attestation fails here (we cannot verify the signature with the
public key given to use by CertifyKey since step 2 does not
set the internal_input_flag so in step 3, when computing CDIs
we do not mix in internal input info. However, in step 4, we would
then incorrectly set context[0]'s internal_input_flag so when we
certify key in step 5 and derive CDIs, the internal input info is
mixed in, leading to a signature verification failure.

Now we make derive child set the internal_input flags on the child
context in order to avoid this problem. I have also added a test for
this scenario.