Return error when attempting to retain default context
According to the spec, "DPE must now allow simultaneous use of a
default context and context handles within the same session: these
are mutually exclusive." If we retain the default context and
generate a new context within the same locality, we violate this.
This commit addresses this behavior and returns an InvalidArgument
error.

Further, we also slightly change the behavior of rotate_context
to align with the spec. If we try to rotate to a default context
and there are other random contexts within the same locality,
we return InvalidArgument since this is also a violation of the
aforementioned part of the spec, since we cannot have both default
and non-default contexts within the same locality.