Move all tests to use shared test file
Make it easier to run DPE tests against different targets by putting all
the test cases in a test vector file. This allows test infrastructure
for a specific target to add required setup and transport code and then
execute the tests in a loop.

In particular, this is to support easily running these tests against a
Caliptra HwModel target.

Additional fixes:

* Make hex string conversion use all caps. This matches what Caliptra
  uses for subject/issuer serialNumber fields.
* Support encoding X.509 Directory Strings as either PrintableString or
  UTF8String. This allows more flexibility when specifying issuer
  encodings.