Refactor logic for `DeriveChild`.
There are quite a few edge cases in `DeriveChild` that need to be
considered. This mainly hitches on if a locality has a default context,
it MUST be the only context in the locality. Which also means, default
and non-default contexts can never be together. This refactors this
logic to be more explicit about when a `DeriveChild` command will be
able to satisfy this requirement.