Exclude root node from certs if it doesn't contain any measurements
There are two ways a context can get measurements - if it got
created in derive child as a child context or if it's tci
was extended. Thus, the root node can only have measurements
if it's tci was extended. Otherwise, we should exclude adding
it's measurements to the certs since they are default buffers
of 0s.

This PR also adds support for extend_tci in the go verification
tests so that we can call it before certifying key. This is necessary
since we do not have derive_child implemented yet in the verification
tests, so we can only ever have the root node to certify. If we skip
adding the root node to the certs, then certify key would fail as there
are no measurements to add to the certs, so we call extend_tci to
create measurements in the root node.