Use the DPE leaf cert issuer in certifyKeyCsr
The CSR parsing library expects the issuer of the CMS content
message to match the issuer of the cert we pass it. The certs
in the cert chain will not have the same issuer since they are
signed by their preceding cert, and not the alias key.

So before passing the cert in the cert chain to the CMS parser,
change its issuer to be the DPE leaf cert's issuer so that
parsing passes.