Destroy dangling retired parent contexts in destroy_ctx
When DestroyContext is called, it is possible that the parents
of the context to be destroyed is a chain of retired contexts.
These should be destroyed, since otherwise it is impossible to
clean them up since their handles are invalid. This could prevent
a potential DOS attack. One caveat is that the retired parent
contexts only get destroyed if they have no other active children.