RustCrypto implementation of DPE crypto.
Adds a RustCrypto implementation as an alternative to openssl.
Openssl is set as the default, and for platform/simulator the two are
mutually exclusive.

For the Cryto trait impl, I generally did not use generics across curves even
though it is possible, because the trait bounds needed were often longer than
the actual shared logic.
There is some shared logic with the openssl impl using the hkdf crate,
which I've moved into a shared module.