Make MAX_HANDLES configurable at runtime
To build DPE with a non-default number of MAX_HANDLES, you can
run ARBITRARY_MAX_HANDLES=10 cargo build --features arbitrary_max_handles.
If the ARBITRARY_MAX_HANDLES env var is not provided or is badly
formatted, the default value of 24 will be used.

Some of the tests will fail for small numbers of max handles,
but this is expected since the tests are intended for a DPE with 24
handles. In the future, we should add tests for DPEs where
MAX_HANDLES=1, 2, etc.

Also, make support features configurable at compile time.