User iterator instead of temporary buffer for CertifyKey TCB
When encoding TcbInfos, do not make a temporary buffer to copy all the
TCI measurements. This adds multiple KiB of stack usage. Instead, add a
function to allow the X.509 library to create an iterator for the TCB of
the node.

This allows the main DPE context array to be used as the backing memory
when encoding TcbInfos instead of using extra stack space.