Don't recompile dpe every single time
The build script always writes to `arbitrary_max_handles.rs`, which
means that a recompile is always required.

This then triggers a rebuild of any project (i.e., Caliptra) that uses
this.

This fixes is so that we only write the file if the contents need to be
changed, so the build can be cached.

Clippy fix