Make the profile in `CommandHdr` dynamic.
This makes `CommandHdr` no longer require the profile cargo feature.
This will now check the profile based on a parameter passed into the
`Command::deserialize` function instead.

Many of the tests now check deserializing commands in both profiles.
Some of these will need to specialized more if the profile affects the
contents of the command. Currently the contents are still hard coded.