Move DPE state into the environment. (#425)
* Move DPE state into the environment.

This commit moves the DPE state out of the DPE instance and into a
separate struct. This allows for more flexibility in how the DPE is
used.

More specifically this will help in Caliptra 2.x to support both an
ECDSA and an ML-DSA DPE instance that share the same DPE state.