Simplify command dispatch.
This removes the `Command` enum completely `deserialize_command` instead
returns a `dyn` reference to the `CommandExecution` trait.

This uses a `dyn` reference instead of an `impl` because `dyn` can take
a performance hit but saves on code size because it is dynamically
dispatched. We are low on memory in Caliptra RT firmware and
we are not concerned with performance.

Dyn references can't have `impl` types in any of the function parameters
so this also switches the `impl DpeTypes` for `DpeEnv` to be a generic.

Also because we can't check equality anymore because the compiler
doesn't know what kind of object it is, the deserialization tests just
check for OK now.