Revert "Convert some TciMeasurement parameters to Digest. (#438)"
This reverts commit d73a83e9880cbedee008a5d3e87292b3613fe654.

Our original plan was to completely remove profile data from build time
variables. This looks to be a bit more involved, especially to make the
TciMeasurement size generic. So we decided in the short term to make the
TCI measurement size static at build time. This makes #444 and #438
obsolete. We may want to re-do them later, but in the meantime it just
makes everything less convenient.